+++ Editor/Processors/SkinnedMeshes/FreezeBlendShapeProcessor.cs
@@ -15,7 +15,16 @@ namespace Anatawa12.AvatarOptimizer.Processors.SkinnedMeshes
 
         public override void Process(OptimizerSession session, MeshInfo2 target, MeshInfo2Holder meshInfo2Holder)
         {
-            var freezeNames = Component.FreezingShapeKeys;
+            FreezeBlendShapes(this, session, target, Component.FreezingShapeKeys);
+        }
+
+        public static void FreezeBlendShapes<TComponent>(
+            EditSkinnedMeshProcessor<TComponent> processor,
+            OptimizerSession session,
+            MeshInfo2 target,
+            HashSet<string> freezeNames
+        ) where TComponent : EditSkinnedMeshComponent
+        {
             var freezes = new BitArray(target.BlendShapes.Count);
             for (var i = 0; i < target.BlendShapes.Count; i++)
                 freezes[i] = freezeNames.Contains(target.BlendShapes[i].name);
@@ -44,7 +53,7 @@ namespace Anatawa12.AvatarOptimizer.Processors.SkinnedMeshes
                     if (!freezes[srcI])
                     {
                         // for keep prop: move the BlendShape index. name is not changed.
-                        session.MappingBuilder.RecordMoveProperty(Target, 
+                        session.MappingBuilder.RecordMoveProperty(processor.Target, 
                             VProp.BlendShapeIndex(srcI), 
                             VProp.BlendShapeIndex(dstI));
                         target.BlendShapes[dstI++] = target.BlendShapes[srcI];
@@ -52,8 +61,8 @@ namespace Anatawa12.AvatarOptimizer.Processors.SkinnedMeshes
                     else
                     {
                         // for frozen prop: remove that BlendShape
-                        session.MappingBuilder.RecordRemoveProperty(Target, VProp.BlendShapeIndex(srcI));
-                        session.MappingBuilder.RecordRemoveProperty(Target, $"blendShape.{target.BlendShapes[srcI].name}");
+                        session.MappingBuilder.RecordRemoveProperty(processor.Target, VProp.BlendShapeIndex(srcI));
+                        session.MappingBuilder.RecordRemoveProperty(processor.Target, $"blendShape.{target.BlendShapes[srcI].name}");
                     }
                 }